Factor assumptions into their own module
