+++ crucible/src/Lang/Crucible/Backend/ProofGoals.hs
@@ -18,6 +18,7 @@ module Lang.Crucible.Backend.ProofGoals
   , traverseGoals, traverseOnlyGoals
   , traverseGoalsWithAssumptions
   , traverseGoalsSeq
+    -- ** Proving goals
 
     -- * Goal collector
   , FrameIdentifier(..), GoalCollector
@@ -44,6 +45,14 @@ import           Data.Functor.Const
 import           Data.Sequence (Seq)
 import qualified Data.Sequence as Seq
 import           Data.Word
+
+import qualified What4.Interface as W4
+import qualified What4.Expr as WE
+import qualified What4.Config as W4C
+import qualified What4.Solver.Yices as W4Y
+import qualified What4.Solver.Adapter as WSA
+import qualified What4.SatResult as W4R
+import qualified What4.Partial as W4
 
 -- | A proof goal consists of a collection of assumptions
 --   that were in scope when an assertion was made, together
